Refactor chess activities for future AI support
